Refactor large test file into separate files
Re #23027
