+++ qt/scientific_interfaces/test/ISISReflectometry/Batch/BatchJobRunnerTest.h
@@ -30,12 +30,10 @@ using testing::NiceMock;
 using testing::Return;
 using testing::_;
 
-class BatchJobRunnerTest : public CxxTest::TestSuite {
+class BatchJobRunnerTest {
 public:
-  // This pair of boilerplate methods prevent the suite being created statically
-  // This means the constructor isn't called when running other tests
-  static BatchJobRunnerTest *createSuite() { return new BatchJobRunnerTest(); }
-  static void destroySuite(BatchJobRunnerTest *suite) { delete suite; }
+  // The boilerplate methods are not included because this is just a base
+  // class
 
   BatchJobRunnerTest()
       : m_instruments{"INTER", "OFFSPEC", "POLREF", "SURF", "CRISP"},
@@ -45,610 +43,7 @@ public:
     m_jobAlgorithm = boost::make_shared<MockBatchJobAlgorithm>();
   }
 
-  void testInitialisedWithNonRunningState() {
-    auto jobRunner = makeJobRunner();
-    TS_ASSERT_EQUALS(jobRunner.isProcessing(), false);
-    TS_ASSERT_EQUALS(jobRunner.isAutoreducing(), false);
-    verifyAndClear();
-  }
-
-  void testReductionResumed() {
-    auto jobRunner = makeJobRunner();
-    jobRunner.reductionResumed();
-    auto const hasSelection = false;
-    TS_ASSERT_EQUALS(jobRunner.isProcessing(), true);
-    TS_ASSERT_EQUALS(jobRunner.isAutoreducing(), false);
-    TS_ASSERT_EQUALS(jobRunner.m_reprocessFailed, hasSelection);
-    TS_ASSERT_EQUALS(jobRunner.m_processAll, !hasSelection);
-    verifyAndClear();
-  }
-
-  void testReductionPaused() {
-    auto jobRunner = makeJobRunner();
-    jobRunner.reductionPaused();
-    TS_ASSERT_EQUALS(jobRunner.isProcessing(), false);
-    verifyAndClear();
-  }
-
-  void testAutoreductionResumed() {
-    auto jobRunner = makeJobRunner();
-    jobRunner.autoreductionResumed();
-    TS_ASSERT_EQUALS(jobRunner.isProcessing(), true);
-    TS_ASSERT_EQUALS(jobRunner.isAutoreducing(), true);
-    TS_ASSERT_EQUALS(jobRunner.m_reprocessFailed, true);
-    TS_ASSERT_EQUALS(jobRunner.m_processAll, true);
-    verifyAndClear();
-  }
-
-  void testAutoreductionPaused() {
-    auto jobRunner = makeJobRunner();
-    jobRunner.autoreductionPaused();
-    TS_ASSERT_EQUALS(jobRunner.isAutoreducing(), false);
-    verifyAndClear();
-  }
-
-  void testSetReprocessFailedItems() {
-    auto jobRunner = makeJobRunner();
-    jobRunner.setReprocessFailedItems(true);
-    TS_ASSERT_EQUALS(jobRunner.m_reprocessFailed, true);
-    verifyAndClear();
-  }
-
-  void testGetAlgorithmsWithEmptyModel() {
-    auto jobRunner = makeJobRunner();
-    auto const algorithms = jobRunner.getAlgorithms();
-    TS_ASSERT_EQUALS(algorithms, std::deque<IConfiguredAlgorithm_sptr>());
-    verifyAndClear();
-  }
-
-  void testGetAlgorithmsWithMultiGroupModel() {
-    // TODO add content to model
-    auto jobRunner = makeJobRunner();
-    auto const algorithms = jobRunner.getAlgorithms();
-    TS_ASSERT_EQUALS(algorithms, std::deque<IConfiguredAlgorithm_sptr>());
-    verifyAndClear();
-  }
-
-  void testAlgorithmStarted() {
-    auto row = makeRow("12345", 0.5);
-    auto jobRunner = makeJobRunner();
-
-    EXPECT_CALL(*m_jobAlgorithm, item())
-        .Times(AtLeast(1))
-        .WillRepeatedly(Return(&row));
-
-    jobRunner.algorithmStarted(m_jobAlgorithm);
-    TS_ASSERT_EQUALS(row.state(), State::ITEM_RUNNING);
-    TS_ASSERT_EQUALS(row.reducedWorkspaceNames().iVsLambda(), "");
-    TS_ASSERT_EQUALS(row.reducedWorkspaceNames().iVsQ(), "");
-    TS_ASSERT_EQUALS(row.reducedWorkspaceNames().iVsQBinned(), "");
-    verifyAndClear();
-  }
-
-  void testAlgorithmComplete() {
-    auto row = makeRow("12345", 0.5);
-    auto jobRunner = makeJobRunner();
-    auto iVsQ = createWorkspace();
-    auto iVsQBin = createWorkspace();
-
-    EXPECT_CALL(*m_jobAlgorithm, item())
-        .Times(AtLeast(1))
-        .WillRepeatedly(Return(&row));
-    EXPECT_CALL(*m_jobAlgorithm, outputWorkspaceNames())
-        .Times(1)
-        .WillOnce(Return(std::vector<std::string>{"", "IvsQ", "IvsQBin"}));
-    EXPECT_CALL(*m_jobAlgorithm, outputWorkspaceNameToWorkspace())
-        .Times(1)
-        .WillOnce(Return(std::map<std::string, Workspace_sptr>{
-            {"OutputWorkspace", iVsQ}, {"OutputWorkspaceBinned", iVsQBin}}));
-
-    jobRunner.algorithmComplete(m_jobAlgorithm);
-    TS_ASSERT_EQUALS(row.state(), State::ITEM_COMPLETE);
-    TS_ASSERT_EQUALS(row.reducedWorkspaceNames().iVsLambda(), "");
-    TS_ASSERT_EQUALS(row.reducedWorkspaceNames().iVsQ(), "IvsQ");
-    TS_ASSERT_EQUALS(row.reducedWorkspaceNames().iVsQBinned(), "IvsQBin");
-    verifyAndClear();
-  }
-
-  void testAlgorithmError() {
-    auto row = makeRow("12345", 0.5);
-    auto jobRunner = makeJobRunner();
-    auto message = std::string("test error message");
-
-    EXPECT_CALL(*m_jobAlgorithm, item())
-        .Times(AtLeast(1))
-        .WillRepeatedly(Return(&row));
-
-    jobRunner.algorithmError(m_jobAlgorithm, message);
-    TS_ASSERT_EQUALS(row.state(), State::ITEM_ERROR);
-    TS_ASSERT_EQUALS(row.message(), message);
-    TS_ASSERT_EQUALS(row.reducedWorkspaceNames().iVsLambda(), "");
-    TS_ASSERT_EQUALS(row.reducedWorkspaceNames().iVsQ(), "");
-    TS_ASSERT_EQUALS(row.reducedWorkspaceNames().iVsQBinned(), "");
-    verifyAndClear();
-  }
-
-  void testGetWorkspacesToSaveForOnlyRowInGroup() {
-    auto jobRunner = makeJobRunner(oneGroupWithARowModel());
-    auto *row = getRow(jobRunner, 0, 0);
-    row->setOutputNames({"", "IvsQ", "IvsQBin"});
-
-    EXPECT_CALL(*m_jobAlgorithm, item())
-        .Times(AtLeast(1))
-        .WillRepeatedly(Return(row));
-
-    // For a single row, we save the binned workspace for the row
-    auto workspacesToSave =
-        jobRunner.algorithmOutputWorkspacesToSave(m_jobAlgorithm);
-    TS_ASSERT_EQUALS(workspacesToSave, std::vector<std::string>{"IvsQBin"});
-
-    verifyAndClear();
-  }
-
-  void testGetWorkspacesToSaveForRowInMultiRowGroup() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    auto *row = getRow(jobRunner, 0, 0);
-    row->setOutputNames({"", "IvsQ", "IvsQBin"});
-
-    EXPECT_CALL(*m_jobAlgorithm, item())
-        .Times(AtLeast(1))
-        .WillRepeatedly(Return(row));
-
-    // For multiple rows, we don't save any workspaces
-    auto workspacesToSave =
-        jobRunner.algorithmOutputWorkspacesToSave(m_jobAlgorithm);
-    TS_ASSERT_EQUALS(workspacesToSave, std::vector<std::string>{});
-
-    verifyAndClear();
-  }
-
-  void testGetWorkspacesToSaveForGroup() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    auto &group = getGroup(jobRunner, 0);
-    group.setOutputNames({
-        "stitched_test",
-    });
-
-    EXPECT_CALL(*m_jobAlgorithm, item())
-        .Times(AtLeast(1))
-        .WillRepeatedly(Return(&group));
-
-    auto workspacesToSave =
-        jobRunner.algorithmOutputWorkspacesToSave(m_jobAlgorithm);
-    TS_ASSERT_EQUALS(workspacesToSave,
-                     std::vector<std::string>{"stitched_test"});
-
-    verifyAndClear();
-  }
-
-  void testDeletedWorkspaceResetsStateForRow() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    auto *row = getRow(jobRunner, 0, 1);
-    row->setSuccess();
-    row->setOutputNames({"", "IvsQ_test", "IvsQBin_test"});
-
-    jobRunner.notifyWorkspaceDeleted("IvsQBin_test");
-    TS_ASSERT_EQUALS(row->state(), State::ITEM_NOT_STARTED);
-    verifyAndClear();
-  }
-
-  void testDeletedWorkspaceResetsOutputNamesForRow() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    auto *row = getRow(jobRunner, 0, 1);
-    row->setSuccess();
-    row->setOutputNames({"", "IvsQ_test", "IvsQBin_test"});
-
-    jobRunner.notifyWorkspaceDeleted("IvsQBin_test");
-    TS_ASSERT_EQUALS(row->reducedWorkspaceNames().iVsLambda(), "");
-    TS_ASSERT_EQUALS(row->reducedWorkspaceNames().iVsQ(), "");
-    TS_ASSERT_EQUALS(row->reducedWorkspaceNames().iVsQBinned(), "");
-    verifyAndClear();
-  }
-
-  void testDeleteWorkspaceResetsStateForGroup() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    auto &group = getGroup(jobRunner, 0);
-    group.setSuccess();
-    group.setOutputNames({"stitched_test"});
-
-    jobRunner.notifyWorkspaceDeleted("stitched_test");
-    TS_ASSERT_EQUALS(group.state(), State::ITEM_NOT_STARTED);
-    verifyAndClear();
-  }
-
-  void testDeleteWorkspaceResetsOutputNamesForGroup() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    auto &group = getGroup(jobRunner, 0);
-    group.setSuccess();
-    group.setOutputNames({"stitched_test"});
-
-    jobRunner.notifyWorkspaceDeleted("stitched_test");
-    TS_ASSERT_EQUALS(group.postprocessedWorkspaceName(), "");
-    verifyAndClear();
-  }
-
-  void testRenameWorkspaceDoesNotResetStateForRow() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    auto *row = getRow(jobRunner, 0, 1);
-    row->setSuccess();
-    row->setOutputNames({"", "IvsQ_test", "IvsQBin_test"});
-
-    jobRunner.notifyWorkspaceRenamed("IvsQBin_test", "IvsQBin_new");
-    TS_ASSERT_EQUALS(row->state(), State::ITEM_COMPLETE);
-    verifyAndClear();
-  }
-
-  void testRenameWorkspaceUpdatesCorrectWorkspaceForRow() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    auto *row = getRow(jobRunner, 0, 1);
-    row->setSuccess();
-    row->setOutputNames({"", "IvsQ_test", "IvsQBin_test"});
-
-    jobRunner.notifyWorkspaceRenamed("IvsQBin_test", "IvsQBin_new");
-    TS_ASSERT_EQUALS(row->reducedWorkspaceNames().iVsLambda(), "");
-    TS_ASSERT_EQUALS(row->reducedWorkspaceNames().iVsQ(), "IvsQ_test");
-    TS_ASSERT_EQUALS(row->reducedWorkspaceNames().iVsQBinned(), "IvsQBin_new");
-    verifyAndClear();
-  }
-
-  void testRenameWorkspaceDoesNotResetStateForGroup() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    auto &group = getGroup(jobRunner, 0);
-    group.setSuccess();
-    group.setOutputNames({"stitched_test"});
-
-    jobRunner.notifyWorkspaceRenamed("stitched_test", "stitched_new");
-    TS_ASSERT_EQUALS(group.state(), State::ITEM_COMPLETE);
-    verifyAndClear();
-  }
-
-  void testRenameWorkspaceUpdatesPostprocessedNameForGroup() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    auto &group = getGroup(jobRunner, 0);
-    group.setSuccess();
-    group.setOutputNames({"stitched_test"});
-
-    jobRunner.notifyWorkspaceRenamed("stitched_test", "stitched_new");
-    TS_ASSERT_EQUALS(group.postprocessedWorkspaceName(), "stitched_new");
-    verifyAndClear();
-  }
-
-  void testDeleteAllWorkspacesResetsStateForRowAndGroup() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    auto *row = getRow(jobRunner, 0, 1);
-    auto &group = getGroup(jobRunner, 0);
-    row->setSuccess();
-    row->setOutputNames({"", "IvsQ_test", "IvsQBin_test"});
-    group.setSuccess();
-    group.setOutputNames({"stitched_test"});
-
-    jobRunner.notifyAllWorkspacesDeleted();
-    TS_ASSERT_EQUALS(row->state(), State::ITEM_NOT_STARTED);
-    TS_ASSERT_EQUALS(group.state(), State::ITEM_NOT_STARTED);
-    verifyAndClear();
-  }
-
-  void testDeleteAllWorkspacesResetsOutputNamesForRowAndGroup() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    auto *row = getRow(jobRunner, 0, 1);
-    auto &group = getGroup(jobRunner, 0);
-    row->setSuccess();
-    row->setOutputNames({"", "IvsQ_test", "IvsQBin_test"});
-    group.setSuccess();
-    group.setOutputNames({"stitched_test"});
-
-    jobRunner.notifyAllWorkspacesDeleted();
-    TS_ASSERT_EQUALS(row->reducedWorkspaceNames().iVsLambda(), "");
-    TS_ASSERT_EQUALS(row->reducedWorkspaceNames().iVsQ(), "");
-    TS_ASSERT_EQUALS(row->reducedWorkspaceNames().iVsQBinned(), "");
-    TS_ASSERT_EQUALS(group.postprocessedWorkspaceName(), "");
-    verifyAndClear();
-  }
-
-  void testProgressWithEmptyTable() {
-    auto jobRunner = makeJobRunner(oneEmptyGroupModel());
-    jobRunner.m_processAll = true;
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressWithEmptyGroup() {
-    auto jobRunner = makeJobRunner(oneEmptyGroupModel());
-    jobRunner.m_processAll = true;
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressWhenRowNotStarted() {
-    auto jobRunner = makeJobRunner(oneGroupWithARowModel());
-    jobRunner.m_processAll = true;
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 0);
-  }
-
-  void testProgressWhenRowStarting() {
-    auto jobRunner = makeJobRunner(oneGroupWithARowModel());
-    jobRunner.m_processAll = true;
-    getRow(jobRunner, 0, 0)->setStarting();
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 0);
-  }
-
-  void testProgressWhenRowRunning() {
-    auto jobRunner = makeJobRunner(oneGroupWithARowModel());
-    jobRunner.m_processAll = true;
-    getRow(jobRunner, 0, 0)->setRunning();
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 0);
-  }
-
-  void testProgressWhenRowComplete() {
-    auto jobRunner = makeJobRunner(oneGroupWithARowModel());
-    jobRunner.m_processAll = true;
-    getRow(jobRunner, 0, 0)->setSuccess();
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressWhenRowFailed() {
-    auto jobRunner = makeJobRunner(oneGroupWithARowModel());
-    jobRunner.m_processAll = true;
-    getRow(jobRunner, 0, 0)->setError("error message");
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressWhenGroupNotStarted() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    jobRunner.m_processAll = true;
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 0);
-  }
-
-  void testProgressWhenGroupStarting() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    jobRunner.m_processAll = true;
-    getGroup(jobRunner, 0).setStarting();
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 0);
-  }
-
-  void testProgressWhenGroupRunning() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    jobRunner.m_processAll = true;
-    getGroup(jobRunner, 0).setRunning();
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 0);
-  }
-
-  void testProgressWhenGroupComplete() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    jobRunner.m_processAll = true;
-    getGroup(jobRunner, 0).setSuccess();
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 33);
-  }
-
-  void testProgressWhenGroupError() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    jobRunner.m_processAll = true;
-    getGroup(jobRunner, 0).setError("error message");
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 33);
-  }
-
-  void testProgressExcludesSingleRowGroup() {
-    // Postprocessing is not applicable to a group if it only has one row, so
-    // in this case the single row is the only item that needs processing and
-    // so we expect 100% when that row is complete
-    auto jobRunner = makeJobRunner(oneGroupWithARowModel());
-    jobRunner.m_processAll = true;
-    getRow(jobRunner, 0, 0)->setSuccess();
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressForTwoRowGroupWithOneRowComplete() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    jobRunner.m_processAll = true;
-    getRow(jobRunner, 0, 0)->setSuccess();
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 33);
-  }
-
-  void testProgressForTwoRowGroupWithTwoRowsComplete() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    jobRunner.m_processAll = true;
-    getRow(jobRunner, 0, 0)->setSuccess();
-    getRow(jobRunner, 0, 1)->setSuccess();
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 66);
-  }
-
-  void testProgressForTwoRowGroupWithEverythingComplete() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    jobRunner.m_processAll = true;
-    getGroup(jobRunner, 0).setSuccess();
-    getRow(jobRunner, 0, 0)->setSuccess();
-    getRow(jobRunner, 0, 1)->setSuccess();
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressForTwoGroupsWithOneGroupComplete() {
-    auto jobRunner = makeJobRunner(twoGroupsWithTwoRowsModel());
-    jobRunner.m_processAll = true;
-    getGroup(jobRunner, 0).setSuccess();
-    getRow(jobRunner, 0, 0)->setSuccess();
-    getRow(jobRunner, 0, 1)->setSuccess();
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 50);
-  }
-
-  void testProgressForTwoGroupsWithBothGroupsComplete() {
-    auto jobRunner = makeJobRunner(twoGroupsWithTwoRowsModel());
-    jobRunner.m_processAll = true;
-    getGroup(jobRunner, 0).setSuccess();
-    getRow(jobRunner, 0, 0)->setSuccess();
-    getRow(jobRunner, 0, 1)->setSuccess();
-    getGroup(jobRunner, 1).setSuccess();
-    getRow(jobRunner, 1, 0)->setSuccess();
-    getRow(jobRunner, 1, 1)->setSuccess();
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressOfSelectionWithEmptyTable() {
-    auto jobRunner = makeJobRunner(oneEmptyGroupModel());
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressOfSelectionWithEmptyGroup() {
-    auto jobRunner = makeJobRunner(oneEmptyGroupModel());
-    selectGroup(jobRunner, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressOfSelectionWhenRowNotStarted() {
-    auto jobRunner = makeJobRunner(oneGroupWithARowModel());
-    selectGroup(jobRunner, 0);
-    selectRow(jobRunner, 0, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 0);
-  }
-
-  void testProgressOfSelectionWhenRowStarting() {
-    auto jobRunner = makeJobRunner(oneGroupWithARowModel());
-    getRow(jobRunner, 0, 0)->setStarting();
-    selectGroup(jobRunner, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 0);
-  }
-
-  void testProgressOfSelectionWhenRowRunning() {
-    auto jobRunner = makeJobRunner(oneGroupWithARowModel());
-    getRow(jobRunner, 0, 0)->setRunning();
-    selectGroup(jobRunner, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 0);
-  }
-
-  void testProgressOfSelectionWhenRowComplete() {
-    auto jobRunner = makeJobRunner(oneGroupWithARowModel());
-    getRow(jobRunner, 0, 0)->setSuccess();
-    selectGroup(jobRunner, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressOfSelectionWhenRowFailed() {
-    auto jobRunner = makeJobRunner(oneGroupWithARowModel());
-    getRow(jobRunner, 0, 0)->setError("error message");
-    selectGroup(jobRunner, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressOfSelectionWhenGroupNotStarted() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    selectGroup(jobRunner, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 0);
-  }
-
-  void testProgressOfSelectionWhenGroupStarting() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    getGroup(jobRunner, 0).setStarting();
-    selectGroup(jobRunner, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 0);
-  }
-
-  void testProgressOfSelectionWhenGroupRunning() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    getGroup(jobRunner, 0).setRunning();
-    selectGroup(jobRunner, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 0);
-  }
-
-  void testProgressOfSelectionWhenGroupComplete() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    getGroup(jobRunner, 0).setSuccess();
-    selectGroup(jobRunner, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 33);
-  }
-
-  void testProgressOfSelectionWhenGroupError() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    getGroup(jobRunner, 0).setError("error message");
-    selectGroup(jobRunner, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 33);
-  }
-
-  void testProgressOfSelectionExcludesSingleRowGroup() {
-    // Postprocessing is not applicable to a group if it only has one row, so
-    // so we expect 100% when that row is complete
-    auto jobRunner = makeJobRunner(oneGroupWithARowModel());
-    jobRunner.m_processAll = true;
-    getRow(jobRunner, 0, 0)->setSuccess();
-    selectGroup(jobRunner, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressOfSelectionForTwoRowGroupWithOneRowComplete() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    getRow(jobRunner, 0, 0)->setSuccess();
-    selectGroup(jobRunner, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 33);
-  }
-
-  void testProgressOfSelectionForTwoRowGroupWithTwoRowsComplete() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    getRow(jobRunner, 0, 0)->setSuccess();
-    getRow(jobRunner, 0, 1)->setSuccess();
-    selectGroup(jobRunner, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 66);
-  }
-
-  void testProgressOfSelectionForTwoRowGroupWithEverythingComplete() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    getGroup(jobRunner, 0).setSuccess();
-    getRow(jobRunner, 0, 0)->setSuccess();
-    getRow(jobRunner, 0, 1)->setSuccess();
-    selectGroup(jobRunner, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressOfSelectionForTwoGroupsWithOneGroupComplete() {
-    auto jobRunner = makeJobRunner(twoGroupsWithTwoRowsModel());
-    getGroup(jobRunner, 0).setSuccess();
-    getRow(jobRunner, 0, 0)->setSuccess();
-    getRow(jobRunner, 0, 1)->setSuccess();
-    selectGroup(jobRunner, 0);
-    selectGroup(jobRunner, 1);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 50);
-  }
-
-  void testProgressOfSelectionWithBothChildAndParentItemsSelected() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    getRow(jobRunner, 0, 1)->setSuccess();
-    // The rows are implicitly selected when we select the group, but make sure
-    // they rows are only counted once if we also select one of the rows
-    selectGroup(jobRunner, 0);
-    selectRow(jobRunner, 0, 0);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 33);
-  }
-
-  void testProgressOfSelectionWithOneRowOutOfTwoSelected() {
-    auto jobRunner = makeJobRunner(oneGroupWithTwoRowsModel());
-    getRow(jobRunner, 0, 1)->setSuccess();
-    selectRow(jobRunner, 0, 1);
-    // the selected row is complete
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressOfSelectionForTwoGroupsWithBothGroupsComplete() {
-    auto jobRunner = makeJobRunner(twoGroupsWithTwoRowsModel());
-    getGroup(jobRunner, 0).setSuccess();
-    getRow(jobRunner, 0, 0)->setSuccess();
-    getRow(jobRunner, 0, 1)->setSuccess();
-    getGroup(jobRunner, 1).setSuccess();
-    getRow(jobRunner, 1, 0)->setSuccess();
-    getRow(jobRunner, 1, 1)->setSuccess();
-    selectGroup(jobRunner, 0);
-    selectGroup(jobRunner, 1);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 100);
-  }
-
-  void testProgressOfSelectionExcludesUnselectedGroups() {
-    auto jobRunner = makeJobRunner(twoGroupsWithTwoRowsModel());
-    // first group is 100% complete
-    getGroup(jobRunner, 0).setSuccess();
-    getRow(jobRunner, 0, 0)->setSuccess();
-    getRow(jobRunner, 0, 1)->setSuccess();
-    // second group is 33% complete
-    getRow(jobRunner, 1, 0)->setSuccess();
-    // select second group only
-    selectGroup(jobRunner, 1);
-    TS_ASSERT_EQUALS(jobRunner.percentComplete(), 33);
-  }
-
-private:
+protected:
   std::vector<std::string> m_instruments;
   double m_tolerance;
   Experiment m_experiment;
@@ -659,6 +54,9 @@ private:
 
   class BatchJobRunnerFriend : public BatchJobRunner {
     friend class BatchJobRunnerTest;
+    friend class BatchJobRunnerProcessingTest;
+    friend class BatchJobRunnerProgressBarTest;
+    friend class BatchJobRunnerWorkspacesTest;
 
   public:
     BatchJobRunnerFriend(Batch batch) : BatchJobRunner(batch) {}
